model/requirements.txt
absl-py==0.9.0
asgiref==3.2.7
astor==0.8.1
astunparse==1.6.3
boto3==1.13.9
botocore==1.16.9
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
Django==3.0.6
djangorestframework==3.11.0
docutils==0.15.2
gast==0.3.3
google-auth==1.15.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.29.0
h5py==2.10.0
idna==2.9
importlib-metadata==1.6.0
jmespath==0.10.0
joblib==0.14.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.2.2
nltk==3.5
numpy==1.18.4
oauthlib==3.1.0
opt-einsum==3.2.1
pandas==1.0.3
protobuf==3.12.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
regex==2020.5.14
requests==2.23.0
requests-file==1.5.1
requests-oauthlib==1.3.0
rsa==4.0
s3transfer==0.3.3
scikit-learn==0.23.1
scipy==1.4.1
six==1.14.0
sqlparse==0.3.1
tensorboard==2.2.0
tensorboard-plugin-wit==1.6.0.post3
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
threadpoolctl==2.0.0
tldextract==2.2.2
tqdm==4.46.0
urllib3==1.25.9
Werkzeug==1.0.1
wrapt==1.12.1
xgboost==1.0.1
zipp==3.1.0
